Code cleanup.
Fixed warnings and added comments to JUnit tests.

+++ Calculator/src/com/cf/Calculator/Calculator.java
@@ -106,6 +106,7 @@ public class Calculator {
 
         if (d == 0)
             throw new IllegalArgumentException();
-        return (double) n/d;
+
+        return n/d;
     }
 }
